+++ plugin/exo/Entity/Misc/Label.php
@@ -5,7 +5,6 @@ namespace UJM\ExoBundle\Entity\Misc;
 use Doctrine\ORM\Mapping as ORM;
 use Ramsey\Uuid\Uuid;
 use UJM\ExoBundle\Entity\QuestionType\MatchQuestion;
-use UJM\ExoBundle\Entity\Misc\MatchAssociation;
 use UJM\ExoBundle\Library\Model\ContentTrait;
 use UJM\ExoBundle\Library\Model\OrderTrait;
 
@@ -33,13 +32,6 @@ class Label
      */
     private $uuid;
 
-    /**
-     * @var ArrayCollection
-     *
-     * @ORM\ManyToMany(targetEntity="UJM\ExoBundle\Entity\Misc\MatchAssociation", mappedBy="labels")
-     */
-    private $associations;
-
     use OrderTrait;
 
     use ContentTrait;
@@ -53,7 +45,6 @@ class Label
     public function __construct()
     {
         $this->uuid = Uuid::uuid4()->toString();
-        $this->associations = new ArrayCollection();
     }
 
     /**
@@ -84,42 +75,6 @@ class Label
     public function setUuid($uuid)
     {
         $this->uuid = $uuid;
-    }
-
-    /**
-     * Get associations.
-     *
-     * @return ArrayCollection
-     */
-    public function getAssociations()
-    {
-        return $this->associations;
-    }
-
-    /**
-     * Add an association.
-     *
-     * @param MatchAssociation $association
-     */
-    public function addAssociation(MatchAssociation $association)
-    {
-        if (!$this->associations->contains($association)) {
-            $this->associations->add($association);
-            $association->addProposal($this);
-        }
-    }
-
-    /**
-     * Remove an association.
-     *
-     * @param MatchAssociation $association
-     */
-    public function removeAssociation(MatchAssociation $association)
-    {
-        if ($this->associations->contains($association)) {
-            $this->associations->removeElement($association);
-            $association->removeProposal($this);
-        }
     }
 
     /**
